+++ src/Models/iTunes/ApplicationPrice.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Atomescrochus\EPF\Models\iTunes;
+
+use Illuminate\Database\Eloquent\Model;
+
+class ApplicationPrice extends Model
+{
+    public $timestamps = false;
+    
+    protected $connection = 'apple-epf';
+    protected $table = 'application_price';
+    protected $primaryKey = "storefront_id";
+    protected $fillable = ['export_date', 'application_id', 'retail_price', 'currency_code', 'storefront_id'];
+}