+++ src/Models/iTunes/Genre.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Atomescrochus\EPF\Models\iTunes;
+
+use Atomescrochus\EPF\Traits\ExportDate;
+use Illuminate\Database\Eloquent\Model;
+
+class Genre extends Model
+{
+    use ExportDate;
+
+    public $timestamps = false;
+    protected $connection = 'apple-epf';
+    protected $table = 'genre';
+    protected $primaryKey = "genre_id";
+    protected $fillable = ['export_date', 'genre_id', 'parent_id', 'name'];
+}