+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Command to import to database:
     - Can download files for full and increment imports
     - Can verify the md5 checksum of downloaded files
+    - Can extract downloade `.tbz` files
 
 ## 0.0.1 (2017-03-18)
 - Base skeleton for the package, based on appstract/php-skeleton