+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Can download files for full and increment imports
     - Can verify the md5 checksum of downloaded files
     - Can extract download `tbz` files
+    - Seems to be able to import
 
 ## 0.0.1 (2017-03-18)
 - Base skeleton for the package, based on appstract/php-skeleton