+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
-## 0.0.1
+## 0.1.0 (2017-03-23)
+- All models are presents.
+
+## 0.0.1 (2017-03-18)
 - Base skeleton for the package, based on appstract/php-skeleton